Add files via upload
This is a browser weather app made with HTML, CSS, and JavaScript. To run this you must use a live server in an IDE. This has my own personal API key, and can stop working at any time. Get your own API key in the open weather website and replace it in the .js file.